MAINT: More code reorganization
